refactor izumi
